requirements.txt
discord.py>=2.4.0,<3.0.0
aiohttp>=3.10.0,<4.0.0
aiosqlite>=0.20.0,<1.0.0
httpx>=0.27.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.7.0,<3.0.0
PyYAML>=6.0.1,<7.0.0
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
